Move tasks from mpci.registration to mpci.tasks
